requirements.txt
# Git integration
gitpython>=3.1.30

# CLI interface
typer>=0.9.0
rich>=13.3.5

# LLM integration
openai>=1.1.0
anthropic>=0.5.0

# Optional local model support
# ollama>=0.1.0
# gpt4all>=1.0.0

# Utility packages
pydantic>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
